# packages/api-contract/openapi.json
{
  "components": {
    "schemas": {
...
      "AirIndex": {
        "description": "European AQI (app/air_index.py, ADR-015). `level` None = \"no index\" (never a\ndefault GOOD); DERIVED, not a Measurement (rule #7).",
        "properties": {
          "complete": {
            "title": "Complete",
            "type": "boolean"
          },
          "dominant": {
            "items": {
              "type": "string"
            },
            "title": "Dominant",
            "type": "array"
          },
          "level": {
            "anyOf": [
              {
                "enum": [
                  "GOOD",
                  "FAIR",
                  "MODERATE",
                  "POOR",
                  "VERY_POOR",
                  "EXTREMELY_POOR"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "missing": {
            "additionalProperties": {
              "enum": [
                "MISSING",
                "STALE",
                "UNIT",
                "INVALID"
              ],
              "type": "string"
            },
            "title": "Missing",
            "type": "object"
          },
          "params": {
            "additionalProperties": {
              "enum": [
                "GOOD",
                "FAIR",
                "MODERATE",
                "POOR",
                "VERY_POOR",
                "EXTREMELY_POOR"
              ],
              "type": "string"
            },
            "title": "Params",
            "type": "object"
          },
          "valid_until": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          }
        },
        "required": [
          "level",
          "complete",
          "params",
          "dominant",
          "missing",
          "valid_until"
        ],
        "title": "AirIndex",
        "type": "object"
      },
      "AirLatestResponse": {
        "properties": {
          "stations": {
            "items": {
              "$ref": "#/components/schemas/AirStation"
            },
            "title": "Stations",
            "type": "array"
          }
        },
        "required": [
          "stations"
        ],
        "title": "AirLatestResponse",
        "type": "object"
      },
      "AirParam": {
        "properties": {
          "freshness": {
            "enum": [
              "FRESH",
              "RECENT",
              "STALE"
            ],
            "title": "Freshness",
            "type": "string"
          },
          "observed_at": {
            "title": "Observed At",
            "type": "string"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "value",
          "unit",
          "observed_at",
          "freshness"
        ],
        "title": "AirParam",
        "type": "object"
      },
      "AirStation": {
        "properties": {
          "assignment_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignment Method"
          },
          "coverage": {
            "anyOf": [
              {
                "enum": [
                  "exact",
                  "nearby",
                  "regional"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage"
          },
          "coverage_radius_km": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage Radius Km"
          },
          "distance_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Km"
          },
          "index": {
            "$ref": "#/components/schemas/AirIndex"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "params": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AirParam"
            },
            "title": "Params",
            "type": "object"
          },
          "source": {
            "const": "gios",
            "title": "Source",
            "type": "string"
          },
          "station_id": {
            "title": "Station Id",
            "type": "string"
          },
          "station_name": {
            "title": "Station Name",
            "type": "string"
          }
        },
        "required": [
          "station_id",
          "station_name",
          "latitude",
          "longitude",
          "params",
          "index",
          "source"
        ],
        "title": "AirStation",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Za Oknem API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/air/latest": {
      "get": {
        "description": "Reads only from our own DB (rule #14) — never calls GIOŚ on request.\nData arrives via `python -m app.connectors.gios.ingest` (manual for now, Phase 4).\n\nTASK-4.1: full MVP param set (PM2.5/PM10/NO2/SO2/O3/CO/C6H6), not just PM2.5 —\neach station now returns a `params` dict keyed by param code instead of a single\ntop-level `pm25` field.\n\nADR-025: `?geo_area_id=N` narrows the list to the station assigned to that area\n(nearest within REGIONAL_MAX_KM = 100 km, with `distance_km`, `assignment_method` and the\nADR-029 `coverage` band); no station in range = empty list (\"brak danych dla obszaru\"),\nunknown area = 404.",
        "operationId": "latest_air_quality_api_v1_air_latest_get",
        "parameters": [
          {
            "in": "query",
            "name": "geo_area_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 2147483647,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Geo Area Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AirLatestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Latest Air Quality"
      }
    },
...
  }
}